Complete overhaul and restructure of the repo
