--- chart.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>World population</title>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.7.3/Chart.min.js"></script>
-    <link rel="stylesheet" type="text/css" href="style.css">
-  </head>
-
-  <body style="background-color:white;">
-    <div class="wrapper">
-      <h1>Sample ChartJS Graph</h1>
-
-      <canvas id="myChart" width="400" height="250"></canvas>
-
-
-      <p> Close this window to go back </p>
-      <input type="button" value="Exit" onclick="self.close()">
-    </div>
-
-    <script src="chart.js"></script>
-  </body>
-</html>
