--- module1.html
@@ -1,117 +1,0 @@
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <title>Primacy</title>
-  <link rel="stylesheet" href="style.css">
-</head>
-
-<body>
-  <div class="container" id="background_settings">
-    <!-- Background Sequences -->
-    <div class="box box2"><h1>BACKGROUND SEQUENCES</h1>
-      <div>
-        <input type="file" id="backgroundseqfilepicker" class="input" accept="">
-      </div>
-
-      <table id="backgroundseqtable" class="input_table">
-      </table>
-    </div>
-
-    <div class="box box1"><h1>PCR SALTS</h1>
-      <!-- PCR salts -->
-      <table class="input_table">
-        <tr>
-          <td>
-            Na
-          </td>
-          <td>
-            K
-          </td>
-          <td>
-            Tris
-          </td>
-          <td>
-            Mg
-          </td>
-          <td>
-            dNTPS
-          </td>
-        </tr>
-        <tr>
-          <td>
-            <input id="pcrsodium" type="text" size="2">
-          </td>
-          <td>
-            <input id="pcrpotassium" type="text" size="2">
-          </td>
-          <td>
-            <input id="pcrtromethamine" type="text" size="2">
-          </td>
-          <td>
-            <input id="pcrmagnesium" type="text" size="2">
-          </td>
-          <td>
-            <input id="pcrdntp" type="text" size="2">
-          </td>
-        </tr>
-      </table>
-    </div>
-  </div>
-
-<div class="container" id="neucleotide_input">
-    <div class="box box1" id="input0">
-            <h1>FASTA SEQUENCE</h1>
-            <table class="input_table" >
-                <tr>
-                    <td>
-                        <input type="file" id="fastaFileSelect" class="input" accept="">
-                    </td>
-                </tr>
-            </table>
-            </br>
-            <hr/>
-            <div class="input">
-                <table>
-                  <tr>
-                    <td style="border-bottom-right-radius: 10px; border-bottom-left-radius: 10px;">
-                      <textarea id="sequenceIdentifier" size="20"></textarea>
-                    </td>
-                  <tr>
-                </table>
-                </br>
-            </div>
-          <table class="input_table">
-            <tr>
-              <td>Please enter start of range: </td>
-              <td><input type="text" id="startRange" size=2></td>
-            </tr>
-            <tr>
-              <td>Please enter end of range: </td>
-              <td><input type="text" id="endRange" size=2></td>
-            </tr>
-          </table>
-          <input type="submit" id="resetButton" value="Reset">
-          <div style="align: center; justify-content: center; max-width: 100%; overflow-x: auto; overflow-y: hidden; background-color: white;">
-            <table style="text-align: center; align: center; justify-content: center;" id="regionPicker">
-              <tr>
-              <tr>
-            </table>
-          </div>
-
-          <input type="submit" id="submitButton" value="Submit" data-tooltip="test">
-
-    </div>
-
-</div>
-
-  <div class="tab_container">
-    <button class="tab tab1" id="module1" style="background-color: #012035;">PRIMER COLLECTION</button>
-    <button class="tab tab2" id="module2">PRIMER SCORING</button>
-    <button class="tab tab3" id="module3">OPTIMIZATION</button>
-    <button class="tab tab4" id="module4">OVERVIEW</button>
-  </div>
-
-</body>
-<script src="module1.js"></script>
-</html>
