+++ package.json
@@ -9,7 +9,8 @@
   "author": "Turan Naimey, Austin Kelly, Chance Nelson, Alex Lacy",
   "license": "BSD-3-Clause",
   "devDependencies": {
-    "electron": "^3.1.2"
+    "electron": "^3.1.3",
+    "electron-react-devtools": "^0.5.3"
   },
   "dependencies": {
     "electron-tabs": "^0.9.4"
