+++ src/html/loading_screen.html
@@ -3,14 +3,14 @@
 <head>
   <meta charset="UTF-8">
   <title>Primacy</title>
-  <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="../css/style.css">
 </head>
 
 <body>
   <div class="container">
     <div class="box box1">
       </br>
-      <img src="img/logo.png" id="loading_logo">
+      <img src="../img/logo.png" id="loading_logo">
       </br></br>
       <h1 style="font-size:24px;">LOADING...</h1>
       <div class="box progress_bar">
@@ -34,5 +34,5 @@
   </div>
 
 </body>
-<script src="loading_screen.js"></script>
+<script src="../js/loading_screen.js"></script>
 </html>
