+++ src/html/module1.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <title>Primacy</title>
-  <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="../css/style.css">
 </head>
 
 <body>
@@ -106,12 +106,12 @@
 </div>
 
   <div class="tab_container">
-    <button class="tab tab1" id="module0" style="background-color: #012035;">PRIMER COLLECTION</button>
-    <button class="tab tab2" id="module1">PRIMER SCORING</button>
-    <button class="tab tab3" id="module2">OPTIMIZATION</button>
-    <button class="tab tab4" id="module3">OVERVIEW</button>
+    <button class="tab tab1" id="module1" style="background-color: #012035;">PRIMER COLLECTION</button>
+    <button class="tab tab2" id="module2">PRIMER SCORING</button>
+    <button class="tab tab3" id="module3">OPTIMIZATION</button>
+    <button class="tab tab4" id="module4">OVERVIEW</button>
   </div>
 
 </body>
-<script src="module0.js"></script>
+<script src="../js/module1.js"></script>
 </html>
