+++ src/html/module2.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <title>Primacy</title>
-  <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="../css/style.css">
 </head>
 
 <body>
@@ -47,5 +47,5 @@
   </div>
 
 </body>
-  <script src="module2.js"></script>
+  <script src="../js/module2.js"></script>
 </html>
