+++ src/html/module3.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <title>Primacy</title>
-  <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="../css/style.css">
   <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
   <script src="chart.js"></script>
 </head>
@@ -147,5 +147,5 @@
   </div>
 
 </body>
-  <script src="module3.js"></script>
+  <script src="../js/module3.js"></script>
 </html>
